+++ src/components/atoms/Button/Button.tsx
@@ -1,6 +1,7 @@
 import React from "react"
 import { ButtonCoreProps } from "./ButtonCore"
 import { PrimaryButton } from "./PrimaryButton"
+import { TextButton } from "./TextButton"
 
 export type ButtonProps = ButtonCoreProps & {
   variant?: "primary" | "secondary" | "text"
@@ -8,8 +9,8 @@ export type ButtonProps = ButtonCoreProps & {
 
 export const Button: React.FC<ButtonProps> = (props) => {
     switch (props.variant) {
-        case "primary":
-            return <PrimaryButton {...props} />
+        case "text":
+            return <TextButton {...props} />
         default:
             return <PrimaryButton {...props} />
     }
